+++ shortcut.ps1
@@ -6,11 +6,11 @@ function Set-ShortcutRunAsAdministrator {
     .DESCRIPTION
         Sets the "Run as administrator" flag on the specified shortcut.
 
-    .PARAMETER ShortcutPath
+    .PARAMETER ShortcutLocation
         The path to the shortcut to set the "Run as administrator" flag on.
 
     .EXAMPLE
-        Set-ShortcutRunAsAdministrator -ShortcutPath "C:\Users\UserName\Desktop\MyShortcut.lnk"
+        Set-ShortcutRunAsAdministrator -ShortcutLocation "C:\Users\UserName\Desktop\MyShortcut.lnk"
 
     .NOTES
         This function is based on the answer at https://stackoverflow.com/a/29002207/2562544.
@@ -18,15 +18,15 @@ function Set-ShortcutRunAsAdministrator {
     [CmdletBinding(SupportsShouldProcess)]
     param (
         [Parameter(Mandatory = $true)]
-        [Alias("Path")]
-        [string] $ShortcutPath
+        [Alias("Location")]
+        [string] $ShortcutLocation
     )
 
-    $bytes = [System.IO.File]::ReadAllBytes($ShortcutPath)
+    $bytes = [System.IO.File]::ReadAllBytes($ShortcutLocation)
     $bytes[0x15] = $bytes[0x15] -bor 0x20 #set byte 21 (0x15) bit 6 (0x20) ON
-    
-    if ($PSCmdlet.ShouldProcess($ShortcutPath, "Set 'Run as administrator' flag")) {
-        [System.IO.File]::WriteAllBytes($ShortcutPath, $bytes)
+
+    if ($PSCmdlet.ShouldProcess($ShortcutLocation, "Set 'Run as administrator' flag")) {
+        [System.IO.File]::WriteAllBytes($ShortcutLocation, $bytes)
     }
 }
 
@@ -38,23 +38,23 @@ function Get-ShortcutIconLocation {
     .DESCRIPTION
         Returns the icon location of the specified shortcut.
 
-    .PARAMETER Path
+    .PARAMETER Location
         The path to the shortcut to read the icon location from.
 
     .OUTPUTS
         string - The icon location of the shortcut.
 
     .EXAMPLE
-        Get-ShortcutIconLocation -Path "C:\Users\UserName\Desktop\MyShortcut.lnk"
+        Get-ShortcutIconLocation -Location "C:\Users\UserName\Desktop\MyShortcut.lnk"
     #>
     [CmdletBinding()]
     param (
         [Parameter(Mandatory = $true)]
-        [string] $Path
+        [string] $Location
     )
 
     $wshShell = New-Object -ComObject WScript.Shell
-    $shortcut = $wshShell.CreateShortcut($Path)
- 
+    $shortcut = $wshShell.CreateShortcut($Location)
+
     return $shortcut.IconLocation
 }